deprecated-python-backend/requirements.txt
Flask==3.0.0
requests
pymongo
dnspython
flask-cors
urllib3
python-dotenv